Works with fixed points, solves Karate with Louvain

+++ test_slmpy_locally.sh
@@ -1,5 +1,7 @@
 set -euo pipefail
+if [ $1 == "build" ]; then
+  rm -rf build
+  python setup.py build
+fi
 
-rm -rf build
-python setup.py build
 PYTHONPATH=$PYTHONPATH:$(pwd)/build/lib.linux-x86_64-3.7 pytest test/
